refactor on bayesian side; almost complete testing
